Title: 두 실수의 곱, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250904/두 실수의 곱/the-product-of-two-real-numbers.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-    int a = 5.26;
-    int b = 8.27;
+    float a = 5.26;
+    float b = 8.27;
 
     cout<< fixed;
     cout.precision(3);
